Renamed layer (#24)
